pyproject.toml
[project]
name = "support-env"
version = "1.0.0"
description = "Customer Support RL Environment for OpenEnv"
readme = "README.md"
authors = [
    {name = "Yash Shinde", email = "yash.shinde@example.com"}
]
requires-python = ">=3.10"
keywords = ["reinforcement-learning", "openenv", "customer-support", "nlp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
    "websockets>=11.0",
    "openenv-core>=0.1.0",
    "gradio>=4.0.0",
    "requests>=2.31.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
    "openai>=1.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
llm = [
    "openai>=1.0.0",
]

[project.scripts]
support-env = "server.app:main"
server = "server.app:main"

[project.urls]
Homepage = "https://huggingface.co/spaces/username/support-env"
Documentation = "https://github.com/username/support-env"
Repository = "https://github.com/username/support-env"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["server*", "baseline*", "frontend*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
ignore = ["E501"]
